requirements.txt
flask
sib_api_v3_sdk
python-dotenv